backend/requirements.txt
Flask
flask-cors
firebase-admin
python-dateutil
gunicorn
